correcao da exclusao das pecas

+++ src/Controllers/PecaController.js
@@ -52,28 +52,19 @@ module.exports={
     async deletar(request,response){
         const cod = request.params.cod;
         const con = await db.conecta();
-        const sql = "DELETE FROM Peca WHERE pec_cod=?";
-        
-        const valor = [cod];
-        const result = await db.manipula(sql,valor);
+        let sql = "SELECT * FROM servicopecas s where pec_cod=?";
+        let valor = [cod];
+        let result = await db.consulta(sql,valor);
+        if(result.data.length===0){
+            sql = "DELETE FROM Peca WHERE pec_cod=?";  
+            valor = [cod];
+            result = await db.manipula(sql,valor);
+        }else{
+            sql = "UPDATE Peca SET pec_status = ? "+
+            "WHERE pec_cod = ?";
+            valor = [false,cod];
+            result = await db.manipula(sql,valor);
+        }
         return response.json(result);
-    },
-    async deletarLogico(request,response){
-        const cod = request.params.cod;
-        const con = await db.conecta();
-        const sql = "UPDATE Peca SET pec_status = ? "+
-                    "WHERE pec_cod = ?";
-        
-        const valor = [false,cod];
-        const result = await db.manipula(sql,valor);
-        return response.json(result);
-    },
-    async listarServicoPecas(request,response){
-        const {pec_cod} = request.params;
-        const con = await db.conecta();
-        const sql = "SELECT * FROM servicopecas s where pec_cod=?";
-        const valor = [pec_cod];
-        const pecas = await db.consulta(sql,valor);
-        return response.json(pecas.data);
     }
 }
